+++ CHANGELOG.md
@@ -7,8 +7,11 @@
 ### Changed
 
 - Create settings directory when it doesn't already exist [#97][97]
+- Pass configuration to tolerance checks so that probe tolerance can access it
+  [#98][98]
 
 [97]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/97
+[98]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/98
 
 ## [1.1.2][] - 2019-03-22